+++ programs/drift/src/controller/insurance.rs
@@ -1,5 +1,5 @@
 use anchor_lang::prelude::*;
-use anchor_spl::token::{Token, TokenAccount};
+use anchor_spl::token_interface::{Mint, TokenAccount, TokenInterface};
 use solana_program::msg;
 
 use crate::controller::spot_balance::{
@@ -633,13 +633,14 @@ pub fn transfer_protocol_insurance_fund_stake(
 }
 
 pub fn attempt_settle_revenue_to_insurance_fund<'info>(
-    spot_market_vault: &Account<'info, TokenAccount>,
-    insurance_fund_vault: &Account<'info, TokenAccount>,
+    spot_market_vault: &InterfaceAccount<'info, TokenAccount>,
+    insurance_fund_vault: &InterfaceAccount<'info, TokenAccount>,
     spot_market: &mut SpotMarket,
     now: i64,
-    token_program: &Program<'info, Token>,
+    token_program: &Interface<'info, TokenInterface>,
     drift_signer: &AccountInfo<'info>,
     state: &State,
+    mint: &Option<InterfaceAccount<'info, Mint>>,
 ) -> Result<()> {
     let valid_revenue_settle_time = if spot_market.insurance_fund.revenue_settle_period > 0 {
         let time_until_next_update = on_the_hour_update(
@@ -680,6 +681,7 @@ pub fn attempt_settle_revenue_to_insurance_fund<'info>(
                 drift_signer,
                 state.signer_nonce,
                 token_amount.cast()?,
+                mint,
             )?;
         }