+++ sdk/src/math/spotMarket.ts
@@ -5,8 +5,12 @@ import {
 	SpotBalanceType,
 	SpotMarketAccount,
 } from '../types';
-import { calculateAssetWeight, calculateLiabilityWeight } from './spotBalance';
-import { MARGIN_PRECISION } from '../constants/numericConstants';
+import {
+	calculateAssetWeight,
+	calculateLiabilityWeight,
+	getTokenAmount,
+} from './spotBalance';
+import { MARGIN_PRECISION, ZERO } from '../constants/numericConstants';
 import { numberToSafeBN } from './utils';
 
 export function castNumberToSpotPrecision(
@@ -53,4 +57,26 @@ export function calculateSpotMarketMarginRatio(
 	}
 
 	return marginRatio;
+}
+
+/**
+ * Returns the maximum remaining deposit that can be made to the spot market. If the maxTokenDeposits on the market is zero then there is no limit and this function will also return zero. (so that needs to be checked)
+ * @param market
+ * @returns
+ */
+export function calculateMaxRemainingDeposit(market: SpotMarketAccount) {
+	const marketMaxTokenDeposits = market.maxTokenDeposits;
+
+	if (marketMaxTokenDeposits.eq(ZERO)) {
+		// If the maxTokenDeposits is set to zero then that means there is no limit. Return the largest number we can to represent infinite available deposit.
+		return ZERO;
+	}
+
+	const totalDepositsTokenAmount = getTokenAmount(
+		market.depositBalance,
+		market,
+		SpotBalanceType.DEPOSIT
+	);
+
+	return marketMaxTokenDeposits.sub(totalDepositsTokenAmount);
 }