+++ sdk/src/types.ts
@@ -117,6 +117,7 @@ export class OracleSource {
 	static readonly PYTH_STABLE_COIN = { pythStableCoin: {} };
 	static readonly PYTH_STABLE_COIN_PULL = { pythStableCoinPull: {} };
 	static readonly Prelaunch = { prelaunch: {} };
+	static readonly SWITCHBOARD_ON_DEMAND = { switchboardOnDemand: {} };
 }
 
 export class OrderType {
@@ -745,6 +746,8 @@ export type SpotMarketAccount = {
 	fuelBoostTaker: number;
 	fuelBoostMaker: number;
 	fuelBoostInsurance: number;
+
+	tokenProgram: number;
 };
 
 export type PoolBalance = {