+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: add switchboard on demand integration ([#1154](https://github.com/drift-labs/protocol-v2/pull/1154))
+- program: add support for token 2022 ([#1125](https://github.com/drift-labs/protocol-v2/pull/1125))
+
 ### Fixes
 
 ### Breaking